@@ -1,4 +1,4 @@
-class AddLastScoredToProjects < ActiveRecord::Migration
+class AddLastScoredToProjects < ActiveRecord::Migration[5.1]
   def change
     add_column :projects, :last_scored, :timestamp
   end
