@@ -1,4 +1,4 @@
-class CreateIssues < ActiveRecord::Migration
+class CreateIssues < ActiveRecord::Migration[5.1]
   def change
     create_table :issues do |t|
       t.integer :github_id, null: false
