@@ -16,8 +16,8 @@ RSpec.describe Api::ProjectsController, type: :controller do
       expect(JSON.parse(response.body)['data'].map { |e| e['id'] }).to eq(%w(123))
     end
 
-    it "meta total_records" do
-      expect(JSON.parse(response.body)['meta']['total_records']).to eq(1)
+    it "meta contains the record count" do
+      expect(JSON.parse(response.body)['meta']['record_count']).to eq(1)
     end
   end
 
@@ -26,7 +26,7 @@ RSpec.describe Api::ProjectsController, type: :controller do
       let!(:project) { create(:project, github_id: 123, owner: 'jack', name: 'foo') }
 
       before do
-        get :show, id: 123
+        get :show, params: { id: 123 }
       end
 
       it "responds successfully with an HTTP 200 status code" do
@@ -56,7 +56,7 @@ RSpec.describe Api::ProjectsController, type: :controller do
 
     context '404' do
       before do
-        get :show, id: 999
+        get :show, params: { id: 999 }
       end
 
       it "404" do
