@@ -27,7 +27,7 @@ RSpec.describe ProjectsController, type: :controller do
 
     describe "show by id" do
       before do
-        get :show, id: project.id
+        get :show, params: { id: project.id }
       end
 
       it "responds successfully with an HTTP 200 status code" do
@@ -46,7 +46,7 @@ RSpec.describe ProjectsController, type: :controller do
 
     describe "show by user/repo" do
       before do
-        get :show, user: project.owner, repo: project.name
+        get :show, params: { user: project.owner, repo: project.name }
       end
 
       it "responds successfully with an HTTP 200 status code" do
@@ -65,19 +65,19 @@ RSpec.describe ProjectsController, type: :controller do
 
     describe "missing repo by user/repo" do
       it "redirects to a prompt screen before adding the repo" do
-        get :show, user: 'foo', repo: 'bar'
+        get :show, params: { user: 'foo', repo: 'bar' }
         expect(response).to redirect_to(project_confirm_path(url: 'foo/bar'))
       end
 
       it "redirects on create action too" do
-        get :create, project: { github_url: 'foo/bar' }
+        get :create, params: { project: { github_url: 'foo/bar' } }
         expect(response).to redirect_to(project_confirm_path(url: 'foo/bar'))
       end
 
       it "doesn't attempt to add the project" do
         expect(Project.all.length).to eq(0)
 
-        get :show, user: 'foo', repo: 'bar'
+        get :show, params: { user: 'foo', repo: 'bar' }
 
         expect(Project.all.length).to eq(0)
       end
@@ -91,7 +91,7 @@ RSpec.describe ProjectsController, type: :controller do
     end
 
     it "displays a confirmation screen for adding a project" do
-      get :confirm, url: 'foo/bar'
+      get :confirm, params: { url: 'foo/bar' }
       expect(response).to have_http_status 200
     end
   end
