@@ -16,7 +16,7 @@ RSpec.describe SearchController, type: :controller do
     it 'searches for the given query and loads the projects into @projects' do
       project = create(:project, name: 'pajamas', owner: 'bananas')
 
-      get :index, q: 'bananas'
+      get :index, params: { q: 'bananas' }
       expect(assigns(:projects)).to include(project)
     end
 
@@ -24,7 +24,7 @@ RSpec.describe SearchController, type: :controller do
       java_project = create(:project, owner: 'bananas', main_language: 'java')
       ruby_project = create(:project, owner: 'bananas', main_language: 'ruby')
 
-      get :index, q: 'bananas', main_language: 'ruby'
+      get :index, params: { q: 'bananas', main_language: 'ruby' }
       expect(assigns(:projects)).to include(ruby_project)
       expect(assigns(:projects)).not_to include(java_project)
     end
